+++ app/src/androidMain/kotlin/com/pavi2410/useCompose/demo/MainActivity.kt
@@ -1,0 +1,14 @@
+package com.pavi2410.useCompose.demo
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            App()
+        }
+    }
+}
