--- app/src/androidMain/kotlin/me/pavi2410/useCompose/demo/MainActivity.kt
@@ -1,14 +1,0 @@
-package me.pavi2410.useCompose.demo
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            App()
-        }
-    }
-}
