+++ app/src/commonMain/kotlin/com/pavi2410/useCompose/demo/screens/Examples.kt
@@ -1,0 +1,13 @@
+package com.pavi2410.useCompose.demo.screens
+
+import com.pavi2410.useCompose.demo.Screen
+
+data class ExampleScreen(
+    val screen: Screen,
+    val title: String
+)
+
+val exampleScreens = listOf(
+    ExampleScreen(Screen.Query, "Query"),
+    ExampleScreen(Screen.Mutation, "Mutation")
+)
