--- app/src/commonMain/kotlin/me/pavi2410/useCompose/demo/screens/Examples.kt
@@ -1,13 +1,0 @@
-package me.pavi2410.useCompose.demo.screens
-
-import me.pavi2410.useCompose.demo.Screen
-
-data class ExampleScreen(
-    val screen: Screen,
-    val title: String
-)
-
-val exampleScreens = listOf(
-    ExampleScreen(Screen.Query, "Query"),
-    ExampleScreen(Screen.Mutation, "Mutation")
-)
