--- app/src/commonMain/kotlin/me/pavi2410/useCompose/demo/screens/MainScreen.kt
@@ -1,43 +1,0 @@
-package me.pavi2410.useCompose.demo.screens
-
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-import me.pavi2410.useCompose.demo.Screen
-
-@OptIn(ExperimentalMaterialApi::class, ExperimentalFoundationApi::class)
-@Composable
-fun MainScreen(
-    onNavigate: (Screen) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    LazyColumn(
-        modifier = modifier
-    ) {
-        items(exampleScreens) { screen ->
-            Card(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp)
-                    .border(Dp.Hairline, Color.Green),
-                onClick = {
-                    onNavigate(screen.screen)
-                }) {
-                Column(Modifier.padding(16.dp)) {
-                    Text(screen.title, Modifier.fillMaxSize())
-                }
-            }
-        }
-    }
-}
