+++ app/src/desktopMain/kotlin/com/pavi2410/useCompose/demo/Main.kt
@@ -1,0 +1,13 @@
+package com.pavi2410.useCompose.demo
+
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.application
+
+fun main() = application {
+    Window(
+        onCloseRequest = ::exitApplication,
+        title = "useCompose Demo"
+    ) {
+        App()
+    }
+}
